[CmakeExample][Geometry] : Add class - Circle
Add class - Circle. It has just one function as
of now - Area(). This function too is empty. This
is because the focus right now is onto writing
CMakeLists.txt and I wanted to write it for multiple
source files. What is inside the source file matters
less for us as of now.